requirements.txt
PyMuPDF==1.23.9
openai==1.9.0
PyYAML==6.0.1
ffmpeg-python==0.2.0
setuptools==63.2.0
python-dotenv==1.0.1
requests==2.31.0